PaintTransformation | separated on part files, added docs
